requirements.txt
asgiref==3.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django>=4.0.7
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.1
psycopg2==2.9.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-dotenv==0.20.0
requests==2.28.1
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.1
